+++ python/aistore/sdk/etl/webserver/fastapi_server.py
@@ -37,7 +37,7 @@ class FastAPIServer(ETLServer):
     Utilizes async/await and threading for optimal request handling.
     """
 
-    def __init__(self, host: str = "0.0.0.0", port: int = 80):
+    def __init__(self, host: str = "0.0.0.0", port: int = 8000):
         super().__init__()
         self.host = host
         self.port = port