+++ python/CHANGELOG.md
@@ -5,12 +5,15 @@ All notable changes to the AIStore Python SDK project are documented in this fil
 We structure this changelog in accordance with [Keep a Changelog](https://keepachangelog.com/) guidelines, and this project follows [Semantic Versioning](https://semver.org/).
 
 ---
-
 ## Unreleased
 
 ### Added
 
 - Add serialization utilities for ETLServer subclasses.
+
+### Changed
+
+- Change default Python ETL web server port from `80` to `8000`.
 
 ## [1.13.8] - 2025-05-16